feat: Add solution for BOJ 24495 - 주사위 네개
This commit adds a solution for the problem "BOJ 24495 - 주사위 네개" to the codebase. The solution uses a combination of functions to determine if a set of four dice can form a non-transitive relationship. It iterates through all possible combinations of dice faces and checks if the non-transitive condition is met. If a valid combination is found, it prints "yes", otherwise it prints "no".