feat: Add solution for BOJ 13334 - 집으로
This commit adds a solution for the problem "BOJ 13334 - 집으로" to the codebase. The solution implements a sweeping algorithm to find the maximum number of houses that can be reached within a given distance from the office. It sorts the data based on the office location and uses a heap to keep track of the reachable houses. The final result is the maximum number of houses in the heap.