feat: Add code for finding the nearest tallest building using a stack
This commit adds code for finding the nearest tallest building in a given list of building heights. The algorithm uses a stack to keep track of the indices of the buildings that are taller than the current building. It calculates the distance to the nearest taller building for each building and stores the results in a new list. The result is printed to the console.